requirements.txt
appdirs==1.4.3
appnope==0.1.0
attrs==19.1.0
autopep8==1.4.4
backcall==0.1.0
black==19.10b0
bleach==3.1.4
click==7.1.1
cycler==0.10.0
DateTime==4.3
decorator==4.4.0
defusedxml==0.6.0
distlib==0.3.0
entrypoints==0.3
et-xmlfile==1.0.1
factor-analyzer==0.3.2
filelock==3.0.12
importlib-metadata==1.5.0
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jdcal==1.4.1
jedi==0.15.1
Jinja2==2.10.1
joblib==0.14.0
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.3.1
jupyter-console==6.0.0
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.5.0
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
jupytext==1.3.3
kiwisolver==1.1.0
lxml==4.4.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
nbconvert==5.6.0
nbformat==4.4.0
notebook==6.0.1
numpy==1.17.1
openpyxl==3.0.0
pandas==0.25.1
pandocfilters==1.4.2
parso==0.5.1
pathspec==0.7.0
patsy==0.5.1
pbr==5.4.4
pexpect==4.7.0
pickleshare==0.7.5
plotly==4.5.4
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.9.1
psycopg2-binary==2.8.4
ptyprocess==0.6.0
pycodestyle==2.5.0
Pygments==2.4.2
PyMySQL==0.9.3
pyparsing==2.4.2
pyrsistent==0.15.4
python-dateutil==2.8.0
pytz==2019.2
PyYAML==5.1.2
pyzmq==18.1.0
qtconsole==4.5.5
regex==2020.2.20
retrying==1.3.3
scikit-learn==0.21.3
scipy==1.3.1
seaborn==0.9.0
Send2Trash==1.5.0
six==1.12.0
statsmodels==0.10.2
stevedore==1.32.0
terminado==0.8.2
testpath==0.4.2
toml==0.10.0
tornado==6.0.3
traitlets==4.3.2
typed-ast==1.4.1
virtualenv==20.0.7
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.4
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.5.1
xlrd==1.2.0
zipp==3.0.0
zope.interface==4.6.0